добавление контроллера
