+++ src/components/board.js
@@ -1,0 +1,7 @@
+import {AbstractComponent} from './abstract-component.js';
+
+class Board extends AbstractComponent {
+  getTemplate() {
+    return `<section class="board container"></section>`;
+  }
+}
